Modified the ranking of the lessons. CSRF has been moved to "Broken Access Control"
git-svn-id: http://webgoat.googlecode.com/svn/trunk@81 4033779f-a91e-0410-96ef-6bf7bf53c507

+++  webgoat/main/project/JavaSource/org/owasp/webgoat/lessons/PathBasedAccessControl.java
@@ -282,12 +282,12 @@ public class PathBasedAccessControl extends LessonAdapter
 	return (instructions);
     }
 
-    private final static Integer DEFAULT_RANKING = new Integer(120);
+    private final static Integer DEFAULT_RANKING = new Integer(115);
 
 
     protected Integer getDefaultRanking()
     {
-	return DEFAULT_RANKING;
+    	return DEFAULT_RANKING;
     }
 
 
@@ -298,6 +298,6 @@ public class PathBasedAccessControl extends LessonAdapter
      */
     public String getTitle()
     {
-	return ("How to Bypass a Path Based Access Control Scheme");
+    	return ("How to Bypass a Path Based Access Control Scheme");
     }
 }
